requirements.txt
grpcio==1.62.1
protobuf==5.26.1
https://github.com/riccardobl/openagents-grpc-proto/releases/download/v0.5.1/openagents_grpc_proto-PYTHON.tar.gz
faiss-cpu==1.8.0
numpy==1.21.2